+++ src/components/datetime/DateTime.js
@@ -5,10 +5,11 @@ import BaseComponent from '../base/Base';
 
 import {
   getDateSetting,
+  offsetDate,
+  formatDate,
   getLocaleDateFormatInfo,
   convertFlatpickrToFormat,
   convertFormatToFlatpickr,
-  convertFormatToMoment,
 } from '../../utils/utils';
 import moment from 'moment';
 
@@ -24,6 +25,8 @@ export default class DateTimeComponent extends BaseComponent {
       enableDate: true,
       enableTime: true,
       defaultDate: '',
+      displayInTimezone: 'viewer',
+      timezone: null,
       datepickerMode: 'day',
       datePicker: {
         showWeeks: true,
@@ -119,6 +122,11 @@ export default class DateTimeComponent extends BaseComponent {
   }
 
   get localeFormat() {
+    const dateFormatInfo = getLocaleDateFormatInfo(this.options.language);
+    this.defaultFormat = this.defaultFormat || {
+      date: dateFormatInfo.dayFirst ? 'd/m/Y ' : 'm/d/Y ',
+      time: 'h:i K'
+    };
     let format = '';
 
     if (this.component.enableDate) {
@@ -134,33 +142,32 @@ export default class DateTimeComponent extends BaseComponent {
 
   get dateTimeFormat() {
     return this.component.useLocaleSettings
-      ? this.localeFormat()
+      ? this.localeFormat
       : convertFormatToFlatpickr(_.get(this.component, 'format', 'yyyy-MM-dd HH:mm a'));
   }
 
-  offset(date) {
+  get timezone() {
+    const timezone = this.component.timezone;
+    if (timezone && timezone.abbr) {
+      return timezone;
+    }
     if (
       (this.component.displayInTimezone === 'submission') &&
       this.root &&
       this.root.hasTimezone
     ) {
       return {
-        date: new Date(date.getTime() + ((this.root.submissionOffset + date.getTimezoneOffset()) * 60000)),
-        timezone: ` (${this.root.submissionTimezone})`
+        offset: this.root.submissionOffset,
+        abbr: this.root.submissionTimezone
       };
     }
-    else if (
-      (this.component.displayInTimezone === 'gmt')
-    ) {
+    if (this.component.displayInTimezone === 'gmt') {
       return {
-        date: new Date(date.getTime() + (date.getTimezoneOffset() * 60000)),
-        timezone: ' (GMT)'
+        offset: 0,
+        abbr: 'GST'
       };
     }
-    return {
-      date,
-      timezone: ` (${this.timezone})`
-    };
+    return null;
   }
 
   get config() {
@@ -187,8 +194,8 @@ export default class DateTimeComponent extends BaseComponent {
       formatDate: (date, format) => {
         // Only format this if this is the altFormat and the form is readOnly.
         if (this.options.readOnly && (format === altFormat)) {
-          const offset = this.offset(date);
-          return `${Flatpickr.formatDate(offset.date, format)}${offset.timezone}`;
+          const offset = offsetDate(date, this.timezone);
+          return `${Flatpickr.formatDate(offset.date, format)}${offset.abbr}`;
         }
 
         return Flatpickr.formatDate(date, format);
@@ -289,10 +296,7 @@ export default class DateTimeComponent extends BaseComponent {
     if (!value) {
       return '';
     }
-    const offset = this.offset(moment(value).toDate());
-    const dateFormat = convertFormatToMoment(_.get(this.component, 'format', 'yyyy-MM-dd HH:mm a'));
-    const formatted = moment(offset.date).format(dateFormat);
-    return `${formatted}${offset.timezone}`;
+    return formatDate(value, _.get(this.component, 'format', 'yyyy-MM-dd HH:mm a'), this.timezone);
   }
 
   setValueAt(index, value) {