+++ src/utils/utils.js
@@ -712,6 +712,55 @@ export function isValidDate(date) {
   return _.isDate(date) && !_.isNaN(date.getDate());
 }
 
+/**
+ * Get the current timezone string.
+ *
+ * @return {string}
+ */
+export function currentTimezone() {
+  if (navigator.languages && navigator.languages.length) {
+    return (new Date()).toLocaleTimeString(navigator.languages[0], {
+      timeZoneName:'short'
+    }).split(' ')[2];
+  }
+  return moment().format('Z');
+}
+
+/**
+ * Get an offset date provided a date object and timezone object.
+ *
+ * @param date
+ * @param timezone
+ * @return {Date}
+ */
+export function offsetDate(date, timezone) {
+  if (!timezone) {
+    return {
+      date,
+      abbr: ` (${currentTimezone()})`
+    };
+  }
+  return {
+    date: new Date(date.getTime() + ((parseInt(timezone.offset, 10) + date.getTimezoneOffset()) * 60000)),
+    abbr: timezone.abbr ? ` (${timezone.abbr})` : ''
+  };
+}
+
+/**
+ * Format a date provided a value, formate, and timezone object.
+ *
+ * @param value
+ * @param format
+ * @param timezone
+ * @return {string}
+ */
+export function formatDate(value, format, timezone) {
+  const date = moment(value).toDate();
+  const offset = offsetDate(date, timezone);
+  const dateFormat = convertFormatToMoment(format);
+  return `${moment(offset.date).format(dateFormat)}${offset.abbr}`;
+}
+
 export function getLocaleDateFormatInfo(locale) {
   const formatInfo = {};
 
@@ -983,4 +1032,38 @@ export function delay(fn, delay = 0, ...args) {
   earlyCall.cancel = cancel;
 
   return earlyCall;
+}
+
+/**
+ * Iterate the given key to make it unique.
+ *
+ * @param {String} key
+ *   Modify the component key to be unique.
+ *
+ * @returns {String}
+ *   The new component key.
+ */
+export function iterateKey(key) {
+  if (!key.match(/(\d+)$/)) {
+    return `${key}2`;
+  }
+
+  return key.replace(/(\d+)$/, function(suffix) {
+    return Number(suffix) + 1;
+  });
+}
+
+/**
+ * Determines a unique key within a map provided the base key.
+ *
+ * @param map
+ * @param base
+ * @return {*}
+ */
+export function uniqueKey(map, base) {
+  let newKey = base;
+  while (map.hasOwnProperty(newKey)) {
+    newKey = iterateKey(newKey);
+  }
+  return newKey;
 }